x86/vioapic: expand hvm_vioapic to contain vIO APIC internal state
This is required in order to have a variable number of vIO APIC pins, instead
of the current fixed value (48). Note that this patch only expands the fields
of the hvm_vioapic struct, without actually introducing any new fields or
functionality.

The reason to expand the hvm_vioapic structure instead of the hvm_hw_vioapic
one is that the variable number of pins functionality is only going to be used
by the hardware domain, so no modifications are needed to the save format.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Cc: Jan Beulich <jbeulich@suse.com>
Cc: Andrew Cooper <andrew.cooper3@citrix.com>
---
Changes since v2:
 - Change patch title.
 - Use an unnamed struct to store the vioapic state inside of hvm_vioapic.
 - Use a define to declare the hvm_hw_vioapic struct (and the equivalent
   unnamed struct inside of hvm_vioapic).
 - Remove the BUILD_BUG_ON.

Changes since v1:
 - New in this version.