x86/hvm: convert gsi_assert_count into a variable size array
Rearrange the fields of hvm_irq so that gsi_assert_count can be converted into
a variable size array and add a new field to account the number of GSIs.

Due to this changes the irq member in the hvm_domain struct also needs to
become a pointer set at runtime.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Cc: Jan Beulich <jbeulich@suse.com>
Cc: Andrew Cooper <andrew.cooper3@citrix.com>
---
Changes since v2:
 - Parenthesize is_hvm_pv_evtchn_domain argument.
 - Add a build BUG_ON to make sure DomU number of IRQs covers the ISA range at
   least.
 - Add an ASSERT to make sure nr_gsis covers the ISA range (those two are
   identical now, but that's going to change when Dom0 introduces a variable
   number of GSIs, hence the ASSERT and the BUILD_BUG_ON).
 - Don't expand the ASSERTs in the irq assert/deassert routines (the above
   additions already cover those).
 - Use %2 as format specifier to print the GSIs assert count (the existing code
   has been left as-is, using %2.2 instead).