libxl: Introduce hack to allow PVH mode to add a shim
libxl will look for LIBXL_PVSHIM_PATH and LIBXL_PVSHIM_CMDLINE
environment variables.  If the first is present, it will boot with the
shim and the existing kernel / ramdisk.  (That is, the shim as the "kernel" and the
kernel and ramdisk both as extra modules.)

If not, it will just boot the kernel / ramdisk directly (that is, with
the kernel as "kernel" and the ramdisk as a module).

Signed-off-by: George Dunlap <george.dunlap@citrix.com>
---
To be replaced with proper toolstack side patches