x86/efi: move the logic to detect PE build support
So that it can be used by the linker script also. With this change the
linker script processor is capable of knowing whether the build has
been performed with PE support or not, and thus create a dummy efi
symbol only when required.

This avoid using a conditional expression with DEFINED on the linker
script, and prevents the following build error with lld (LLVM linker):

ld    -melf_x86_64_fbsd  -T xen.lds -N prelink.o --build-id=sha1 \
    /root/src/xen/xen/common/symbols-dummy.o -o /root/src/xen/xen/.xen-syms.0
ld: error: xen.lds:233: symbol not found: efi

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Cc: Jan Beulich <jbeulich@suse.com>
Cc: Andrew Cooper <andrew.cooper3@citrix.com>
Cc: Daniel Kiper <daniel.kiper@oracle.com>