vpci: add tear down functions
Tear down functions are not mandatory. Note that this patch just
implements the framework, but doesn't implement any tear down function
yet.

The list of vpci init and teardown functions is also moved outside of
the init section. This is in preparation for SR-IOV support, which
requires the ability to add and remove pci devices (and vpci handlers)
after boot.

No functional change intended.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Cc: Andrew Cooper <andrew.cooper3@citrix.com>
Cc: George Dunlap <George.Dunlap@eu.citrix.com>
Cc: Ian Jackson <ian.jackson@eu.citrix.com>
Cc: Jan Beulich <jbeulich@suse.com>
Cc: Julien Grall <julien.grall@arm.com>
Cc: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Cc: Stefano Stabellini <sstabellini@kernel.org>
Cc: Tim Deegan <tim@xen.org>
Cc: Wei Liu <wei.liu2@citrix.com>
---
Changes since v1:
 - Remove a redundant vpci != NULL check.
 - Expand commit message to mention moving the vpci handlers out of
   the init section.